requirements.txt
arango-orm==0.5.7
bcrypt==3.1.6
Bootstrap-Flask==1.0.10
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==1.0.3
Flask-Bcrypt==0.7.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
marshmallow==2.16.0
mimesis==3.2.0
neobolt==1.7.13
neotime==1.7.4
numpy==1.16.4
pandas==0.24.2
passlib==1.7.1
prompt-toolkit==2.0.9
py2neo==4.3.0
pycparser==2.19
Pygments==2.3.1
pyorient==1.5.5
python-arango==4.4.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
sh==1.12.14
six==1.12.0
urllib3==1.24.3
wcwidth==0.1.7
Werkzeug==0.15.4
